Separate queues for handling large and small files with folders

+++ src/main/java/com/yet/dsync/service/DownloadService.java
@@ -16,25 +16,31 @@ package com.yet.dsync.service;
 
 import java.io.File;
 import java.io.FileOutputStream;
+import java.io.IOException;
 import java.util.Collection;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.PriorityBlockingQueue;
 
+import com.dropbox.core.v2.files.DownloadErrorException;
 import com.yet.dsync.dao.MetadataDao;
 import com.yet.dsync.dto.FileData;
 import com.yet.dsync.exception.DSyncClientException;
 
 public class DownloadService {
     
-    private static final int THREAD_NUMBER = 5;
+    private static final int QUICK_THREAD_NUMBER = 5;
+    private static final int SLOW_THREAD_NUMBER = 2;
+    
+    private static final long SLOW_THRESHOLD = 256*1024; //256KB
     
     private final MetadataDao metadaDao;
     private final LocalFolderService localFolderService;
     private final DropboxService dropboxService;
     
-    private final BlockingQueue<FileData> downloadQueue;
+    private final BlockingQueue<FileData> quickDownloadQueue;
+    private final BlockingQueue<FileData> slowDownloadQueue;
     
     private final ExecutorService executorService;
     
@@ -43,8 +49,10 @@ public class DownloadService {
         this.localFolderService = localFolderService;
         this.dropboxService = dropboxService;
         
-        this.downloadQueue = createDownloadQueue();
-        this.executorService = Executors.newFixedThreadPool(THREAD_NUMBER);
+        this.slowDownloadQueue = createDownloadQueue();
+        this.quickDownloadQueue = createDownloadQueue();
+        
+        this.executorService = Executors.newFixedThreadPool(SLOW_THREAD_NUMBER + QUICK_THREAD_NUMBER);
         
         initDownloadThreads();
     }
@@ -72,17 +80,12 @@ public class DownloadService {
     }
     
     private void initDownloadThreads() {
-        for (int i=0; i < THREAD_NUMBER; i++) {
-            executorService.submit(() -> {
-                while (!Thread.interrupted()) {
-                    try {
-                        FileData fileData = downloadQueue.take();
-                        downloadData(fileData);
-                    } catch (Exception e) {
-                        System.err.println("Failed to download file: " + e.getMessage());
-                    }
-                }
-            });
+        for (int i=0; i < QUICK_THREAD_NUMBER; i++) {
+            executorService.submit(new DownloadThread(quickDownloadQueue));
+        }
+        
+        for (int i=0; i < SLOW_THREAD_NUMBER; i++) {
+            executorService.submit(new DownloadThread(slowDownloadQueue));
         }
     }
     
@@ -90,35 +93,76 @@ public class DownloadService {
         if (fileData.getRev() == null) {
             localFolderService.createFolder(fileData.getPathDisplay());
             System.out.println("LOCAL_DIR " + fileData.getPathDisplay());
+            metadaDao.writeLoadedFlag(fileData.getId(), true);
         } else {
             File file = localFolderService.buildFileObject(fileData.getPathDisplay());
-            try (FileOutputStream fos = new FileOutputStream(file)) {
-                dropboxService.downloadFile(fileData.getPathDisplay(), fos);
-            } catch (Exception e) {
-                throw new DSyncClientException(e);
+            
+            if ( file.getParentFile().exists() ) {
+                try (FileOutputStream fos = new FileOutputStream(file)) {
+                    dropboxService.downloadFile(fileData.getPathDisplay(), fos);
+                    metadaDao.writeLoadedFlag(fileData.getId(), true);
+                } catch (IOException e) {
+                    throw new DSyncClientException(e);
+                }
+                System.out.println("DOWNLOADED " + fileData.getPathDisplay());
+            } else {
+                System.out.println("SKIP " + fileData.getPathDisplay());
             }
-            System.out.println("DOWNLOADED " + fileData.getPathDisplay());
         }
-        metadaDao.writeLoadedFlag(fileData.getId(), true);
     }
     
     public void downloadAllNotLoaded() {
         Collection<FileData> allNotLoaded = metadaDao.readAllNotLoaded();
-        allNotLoaded.forEach(f -> {
-            try {
-                downloadQueue.put(f);
-            } catch (Exception e) {
-                throw new DSyncClientException(e);
-            }
-        });
+        System.out.println("Downloading " + allNotLoaded.size() + " objects that are not loaded.");
+        allNotLoaded.forEach(this::scheduleDownload);
     }
     
     public void scheduleDownload(FileData fileData) {
         try {
-            downloadQueue.put(fileData);
+            if (fileData.isFile()) {
+                long size = fileData.getSize();
+                if (size > SLOW_THRESHOLD) {
+                    slowDownloadQueue.put(fileData);
+                } else {
+                    quickDownloadQueue.put(fileData);
+                }
+            } else {
+                quickDownloadQueue.put(fileData);
+            }
         } catch (Exception e) {
             throw new DSyncClientException(e);
         }
+    }
+    
+    private class DownloadThread implements Runnable {
+
+        private BlockingQueue<FileData> queue;
+        
+        public DownloadThread(BlockingQueue<FileData> queue) {
+            this.queue = queue;
+        }
+
+        @Override
+        public void run() {
+            while (!Thread.interrupted()) {
+                try {
+                    FileData fileData = queue.take();
+                    try {
+                        downloadData(fileData);
+                    } catch (DSyncClientException dee) {
+                        if (dee.getCause() instanceof DownloadErrorException) {
+                            metadaDao.deleteByPath(fileData.getPathDisplay());
+                        } else {
+                            throw dee;
+                        }
+                    }
+                    
+                } catch (Exception e) {
+                    System.err.println("Failed to download file: " + e.getMessage());
+                }
+            }
+        }
+        
     }
 
 }
